Removed main.py and cut the project into a multi file system for better parallel workability.
